Refactor spin wheel html and javascript files
The old index.html file has been deleted and replaced with a revised version with updated doctype, meta tags and improved the code structure. The css and javascript files have been renamed which in turn required an update to the file references in index.html. Additionally, we have added a new vcs.xml file in .idea directory to set up Git as our version control system.